+++ nrich-form-configuration-spring-boot-starter/src/test/java/net/croz/nrich/formconfiguration/starter/configuration/NrichFormConfigurationAutoConfigurationTest.java
@@ -20,9 +20,9 @@ package net.croz.nrich.formconfiguration.starter.configuration;
 import net.croz.nrich.formconfiguration.api.customizer.FormConfigurationMappingCustomizer;
 import net.croz.nrich.formconfiguration.api.model.FormConfiguration;
 import net.croz.nrich.formconfiguration.api.service.ConstrainedPropertyValidatorConverterService;
+import net.croz.nrich.formconfiguration.api.service.FieldErrorMessageResolverService;
 import net.croz.nrich.formconfiguration.api.service.FormConfigurationService;
 import net.croz.nrich.formconfiguration.controller.FormConfigurationController;
-import net.croz.nrich.formconfiguration.service.FieldErrorMessageResolverService;
 import net.croz.nrich.formconfiguration.starter.configuration.stub.FormConfigurationTestRequest;
 import net.croz.nrich.formconfiguration.starter.properties.NrichFormConfigurationProperties;
 import net.croz.nrich.javascript.api.converter.JavaToJavascriptTypeConverter;
